Rename alias module
