+++ src/tlc_tools/alias/__init__.py
@@ -1,6 +1,6 @@
 """Tool for managing URL aliases in 3LC objects."""
 
 from .find_aliases import list_aliases
-from .replace_aliases import replace_aliases
+from .replace import replace_aliases
 
 __all__ = ["list_aliases", "replace_aliases"]
