+++ src/tlc_tools/cli/commands/alias.py
@@ -8,7 +8,7 @@ import tlc
 
 from tlc_tools.alias.common import get_input_object
 from tlc_tools.alias.find_aliases import list_aliases
-from tlc_tools.alias.replace_aliases import replace_aliases
+from tlc_tools.alias.replace import replace_aliases
 from tlc_tools.cli import register_tool
 from tlc_tools.cli.logging import setup_logging
